+++ api/src/presentation/containers.py
@@ -6,8 +6,8 @@ from dependency_injector.providers import Factory
 from src.application.film_record.command.interface import IFilmRecordCommandApplication
 from src.application.film_record.query.interface import IFilmRecordQueryApplication
 from src.application.jumanpp.jumanpp_application import JumanppApplication
-from src.presentation.film_record.command.film_record_command_container import FilmRecordCommandContainer
-from src.presentation.film_record.query.film_record_query_container import FilmRecordQueryContainer
+from src.presentation.film_record.command import FilmRecordCommandContainer
+from src.presentation.film_record.query import FilmRecordQueryContainer
 from src.presentation.jumanpp.jumanpp_container import JumanppContainer
