+++ api/src/presentation/film_record/command/__init__.py
@@ -1,0 +1,1 @@
+from .film_record_command_container import FilmRecordCommandContainer
