+++ api/src/presentation/film_record/command/film_record_command_route.py
@@ -1,11 +1,14 @@
-from base_model.film_record.create.film_record_create_model import FilmRecordCreateModel
+# 外部ライブラリ
 from dependency_injector.wiring import Provide, inject
 from fastapi import APIRouter, Depends, status
-from src.application.film_record.command.interface import IFilmRecordCommandApplication
-from src.presentation.containers import Container
-from src.utils.setup_logger import setup_logger
 
 film_record_command_router = APIRouter()
+
+# 独自ライブラリ
+from base_model.film_record.create.film_record_create_model import FilmRecordCreateModel
+from src.application.film_record.command.interface import IFilmRecordCommandApplication
+from src.presentation import Container
+from src.utils.setup_logger import setup_logger
 
 
 @film_record_command_router.post(
