+++ api/src/presentation/film_record/query/__init__.py
@@ -1,0 +1,1 @@
+from .film_record_query_container import FilmRecordQueryContainer
