+++ api/src/presentation/film_record/query/film_record_query_route.py
@@ -2,12 +2,12 @@
 from dependency_injector.wiring import Provide, inject
 from fastapi import APIRouter, Depends
 
+film_record_query_router = APIRouter()
+
 # 独自ライブラリ
 from src.application.film_record.query.interface import IFilmRecordQueryApplication
-from src.presentation.containers import Container
+from src.presentation import Container
 from src.utils.setup_logger import setup_logger
-
-film_record_query_router = APIRouter()
 
 
 @film_record_query_router.get("/{film_record_id}")
